Refactor the TCP_IP module
